Fixed src/

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
     public double calculate(double weight, double height) {
 
-        int bmi = (int) (weight / (height * height));
+        double bmi = (weight / (height * height));
         return bmi;
 
     }
